# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Conveer API
  description: |
    API для управления автоматизированным конвейером создания и прогрева аккаунтов 
    в социальных сетях (VK, Telegram, Mail.ru, Max).
    
    ## Аутентификация
    API использует Bearer токен для аутентификации. Токен можно получить через Telegram бота.
    
    ## Rate Limiting
    - 100 запросов в минуту для обычных endpoint'ов
    - 10 запросов в минуту для операций создания аккаунтов
    
    ## Коды ошибок
    - `400` - Некорректный запрос
    - `401` - Не авторизован
    - `403` - Доступ запрещён
    - `404` - Ресурс не найден
    - `429` - Превышен лимит запросов
    - `500` - Внутренняя ошибка сервера
  version: 1.0.0
  contact:
    name: Conveer Team
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://api.conveer.example.com/v1
    description: Production server
...
paths:
...
  /accounts:
    get:
      tags:
        - accounts
      summary: Получить список аккаунтов
      description: Возвращает список аккаунтов с фильтрацией и пагинацией
      operationId: listAccounts
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: platform
          in: query
          description: Фильтр по платформе
          schema:
            $ref: '#/components/schemas/Platform'
        - name: status
          in: query
          description: Фильтр по статусу
          schema:
            $ref: '#/components/schemas/AccountStatus'
      responses:
        '200':
          description: Список аккаунтов
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Account'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags:
        - accounts
      summary: Создать новый аккаунт
      description: |
        Запускает процесс создания нового аккаунта на указанной платформе.
        Процесс включает получение прокси, SMS-кода и регистрацию.
      operationId: createAccount
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccountRequest'
      responses:
        '202':
          description: Задача на создание аккаунта принята
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateAccountResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /accounts/{accountId}:
    get:
      tags:
        - accounts
      summary: Получить информацию об аккаунте
      operationId: getAccount
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AccountIdParam'
      responses:
        '200':
          description: Информация об аккаунте
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - accounts
      summary: Удалить аккаунт
      description: Удаляет аккаунт и освобождает связанные ресурсы (прокси)
      operationId: deleteAccount
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AccountIdParam'
      responses:
        '204':
          description: Аккаунт удалён
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /proxies:
    get:
      tags:
        - proxies
      summary: Получить список прокси
      operationId: listProxies
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ProxyStatus'
        - name: provider
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Список прокси
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Proxy'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /proxies/allocate:
    post:
      tags:
        - proxies
      summary: Аллоцировать прокси для аккаунта
      description: |
        Выделяет прокси для указанного аккаунта. Если свободных прокси нет,
        покупает новый у провайдера.
      operationId: allocateProxy
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AllocateProxyRequest'
      responses:
        '200':
          description: Прокси выделен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Proxy'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          description: Нет доступных прокси
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /proxies/{proxyId}/release:
    post:
      tags:
        - proxies
      summary: Освободить прокси
      operationId: releaseProxy
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProxyIdParam'
      responses:
        '200':
          description: Прокси освобождён
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /proxies/{proxyId}/rotate:
    post:
      tags:
        - proxies
      summary: Принудительная ротация прокси
      description: Заменяет текущий прокси на новый
      operationId: rotateProxy
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ProxyIdParam'
      responses:
        '200':
          description: Прокси заменён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Proxy'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT токен, полученный через Telegram бота
  parameters:
    PageParam:
      name: page
      in: query
      description: Номер страницы
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Количество элементов на странице
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    AccountIdParam:
      name: accountId
      in: path
      required: true
      description: ID аккаунта
      schema:
        type: string
        pattern: '^[a-f0-9]{24}$'
    ProxyIdParam:
      name: proxyId
      in: path
      required: true
      description: ID прокси
      schema:
        type: string
        pattern: '^[a-f0-9]{24}$'
...
  responses:
    BadRequest:
      description: Некорректный запрос
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Не авторизован
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "UNAUTHORIZED"
            message: "Invalid or expired token"
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    TooManyRequests:
      description: Превышен лимит запросов
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "RATE_LIMIT_EXCEEDED"
            message: "Too many requests, please try again later"
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
    InternalError:
      description: Внутренняя ошибка сервера
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Код ошибки
        message:
          type: string
          description: Описание ошибки
        details:
          type: object
          additionalProperties: true
          description: Дополнительные детали ошибки
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
...
    Platform:
      type: string
      enum: [vk, telegram, mail, max]
      description: Поддерживаемые платформы

    # ==================== Account Schemas ====================
    AccountStatus:
      type: string
      enum:
        - creating
        - active
        - warming
        - banned
        - suspended
        - deleted
      description: Статус аккаунта
    Account:
      type: object
      properties:
        id:
          type: string
          description: MongoDB ObjectID
        platform:
          $ref: '#/components/schemas/Platform'
        username:
          type: string
          description: Имя пользователя на платформе
        phone:
          type: string
          description: Номер телефона (маскированный)
        status:
          $ref: '#/components/schemas/AccountStatus'
        proxy_id:
          type: string
          description: ID привязанного прокси
        warming_status:
          type: string
          enum: [not_started, in_progress, completed, paused]
        warming_progress:
          type: integer
          minimum: 0
          maximum: 100
          description: Прогресс прогрева в процентах
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    CreateAccountRequest:
      type: object
      required:
        - platform
      properties:
        platform:
          $ref: '#/components/schemas/Platform'
        count:
          type: integer
          minimum: 1
          maximum: 10
          default: 1
          description: Количество аккаунтов для создания
        start_warming:
          type: boolean
          default: false
          description: Запустить прогрев сразу после создания
        warming_scenario:
          type: string
          description: ID сценария прогрева (если start_warming=true)
    CreateAccountResponse:
      type: object
      properties:
        task_id:
          type: string
          description: ID задачи создания
        status:
          type: string
          enum: [pending, in_progress]
        estimated_time_seconds:
          type: integer
          description: Ориентировочное время выполнения

    # ==================== Proxy Schemas ====================
    ProxyStatus:
      type: string
      enum:
        - available
        - allocated
        - checking
        - banned
        - expired
      description: Статус прокси
    Proxy:
      type: object
      properties:
        id:
          type: string
        provider:
          type: string
          description: Название провайдера
        host:
          type: string
        port:
          type: integer
        protocol:
          type: string
          enum: [http, https, socks5]
        country:
          type: string
          description: Код страны (ISO 3166-1 alpha-2)
        status:
          $ref: '#/components/schemas/ProxyStatus'
        account_id:
          type: string
          nullable: true
          description: ID аккаунта (если аллоцирован)
        health:
          $ref: '#/components/schemas/ProxyHealth'
        expires_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    ProxyHealth:
      type: object
      properties:
        is_available:
          type: boolean
        latency_ms:
          type: integer
        fraud_score:
          type: number
          format: float
          minimum: 0
          maximum: 100
        is_vpn:
          type: boolean
        last_check:
          type: string
          format: date-time
    AllocateProxyRequest:
      type: object
      required:
        - account_id
      properties:
        account_id:
          type: string
        country:
          type: string
          description: Предпочтительная страна
        provider:
          type: string
          description: Предпочтительный провайдер

    # ==================== Warming Schemas ====================
...